feat(nexus-queue)!: drop the HTTP kicker (no consumers left)
With the TS client publishing straight to the broker and the v1
taskiq/Redis runtime gone, the /enqueue facade had no consumers — remove
it until a producer that can't speak NATS actually appears. kicker.py
becomes probes.py, keeping only create_probes_app (/health, /ready,
/metrics); run_nats_worker loses enable_kicker; RuntimeConfig loses
internal_secret and public_paths (they only fed the kicker auth gate).

BREAKING CHANGE: create_nats_kicker, run_nats_worker(enable_kicker=...)
and RuntimeConfig.{internal_secret,public_paths} are removed. The
re-introduction contract stays recorded in the spec (§8.3).

Docs condensed to the current runtime while at it: the README still
described taskiq + Redis Streams, and spec §8.2/§14 still claimed the TS
client published through the kicker and that conformance ran on
TRANSPORTS = [redis, nats].
