+++ backend/nexus-queue/README.md
@@ -1,27 +1,29 @@
 # nexus-queue
 
-Portable worker runtime for the **Nexus-Queue** standard: taskiq + Redis Streams,
+Portable worker runtime for the **Nexus-Queue** standard: NATS JetStream,
 domain-agnostic, ports-and-adapters. One project builds queues the same way as
 the next, and a worker's handlers move between projects unchanged.
 
 What this package owns (the parts that are the *same* across projects):
 
-- **Namespaced streams** — `nq:{project}:{queue}` (+ `:cg`, `:dlq`) and
-  `nq:{project}:status`, so multiple projects/queues coexist on one Redis
-  (the default global `"taskiq"` stream is never used).
+- **Namespaced subjects/streams** — `nq.{project}.{queue}` on stream
+  `NQ_{PROJECT}_{QUEUE}` (+ DLQ and advisory streams), so multiple
+  projects/queues coexist on one NATS (a transport's default/global queue
+  name is never used).
 - **Versioned envelope** — standard labels (`nq_v`, `nq_task`, `nq_tenant`,
-  `nq_idem`, `nq_trace`, `nq_enqueued_at`, `nq_priority`) on top of taskiq's
-  message, plus typed pydantic payloads.
+  `nq_idem`, `nq_trace`, `nq_enqueued_at`, `nq_priority`) plus typed pydantic
+  payloads.
 - **Ports** — `JobStatePort`, `BlobStorePort`, `IndexPort`, `StatusEventPort`.
   Handlers depend only on these; each project supplies the adapters
   (e.g. Payload vs Postgres/MinIO).
-- **Middleware stack** (broker-level): idempotency (dedup on `nq_idem`),
-  DLQ (dead-letter on retry-exhaustion instead of silent drop), retries with
-  exponential backoff + jitter, OTel tracing (`nq_trace` propagation), and
-  Prometheus metrics.
-- **Producer + kicker** — a Python `Publisher` and a generic HTTP kicker for
-  non-Python producers. The TypeScript producer client ships as
-  `@zetesis/nexus-queue`.
+- **Runtime semantics** — idempotency claims on NATS KV (dedup on `nq_idem`),
+  retries with exponential backoff + jitter, DLQ on retry exhaustion (durable
+  `MAX_DELIVERIES` advisory capture — never a silent drop), OTel tracing
+  (`nq_trace` propagation), and Prometheus metrics.
+- **Producer** — a Python `NatsPublisher`; the TypeScript producer client
+  ships as `@zetesis/nexus-queue`. Both publish to NATS directly — there is
+  no HTTP enqueue facade; the worker's HTTP surface is probes/metrics only
+  (`create_probes_app`).
 
 Spec: `nexus-queue-spec.md`.
