+++ docs/architecture/nexus-queue-jetstream-migration.md
@@ -3,6 +3,7 @@
 > **Estado**: propuesta consolidada (2026-07-03), pendiente de ratificar.
 > **Enmienda 2026-07-08**: el cero-Redis por proyecto queda declarado como estado final con fase propia y diferida (D15 nueva, M10 nueva) — la convivencia Redis+NATS es un estado intermedio de la obra, no el destino.
 > **Enmienda 2026-07-09 (big-bang ZP)**: ZP ejecuta el endgame cero-Redis **directamente en el devcontainer** — M10 implementada (store KV + selector `idempotency_backend`, suite parametrizada por store, 34 asserts verdes en la matriz redis/nats/nats+kv) y validada e2e con Redis apagado. La secuencia dual-consume/drain (M4/M5) deja de aplicar a ZP: el patrón D9 (job en `CREATED` + sweep) hace la cola re-derivable desde Payload, el productor habla HTTP con el kicker, y el tráfico es esporádico — el paso a Kubernetes será una instalación NATS limpia + re-kick de docs pendientes, no un cutover con drain. La maquinaria dual-consume queda documentada para flotas con tráfico en vuelo (nixon), y el código v1 (taskiq/Redis) permanece en el paquete hasta que nixon migre.
+> **Enmienda 2026-07-10 (retirada del kicker)**: tras pasar el cliente TS a publicar directo al broker (y con el v1 taskiq/Redis ya eliminado del paquete), la fachada HTTP `/enqueue` (kicker) se quedó sin consumidores y se retira — el HTTP del worker queda en probes/métricas (`create_probes_app`). Si aparece un productor que no pueda hablar NATS (webhook, serverless), se reintroduce con el contrato registrado en el spec §8.3.
 > **Evaluación de realismo ejecutada** (mismo día): banco reproducible en `backend/nexus-queue/spike-jetstream/` (veredicto completo en su `EVALUATION.md`). Todas las claims a nivel de broker verificadas en vivo (E2–E7 PASS, incluida la central de scale-to-zero); un supuesto de implementación resultó falso y queda enmendado aquí (D7 resuelta, D14 nueva, M1 ejecutada, M3 redefinida).
 > **Fuentes**: informes Nexus-Queue 1/3, 2/3, 3/3 y anexo "Si migrásemos de Redis" + discusión posterior + spike.
 > **Decisión central**: reemplazar Redis Streams por NATS JetStream como transporte de cola **antes** de construir la capa de plataforma (chart, KEDA, CLI), y evolucionar el estándar hacia un modelo contract-first ("OpenAPI de colas") con un **runtime v2 sin taskiq** (D14).
@@ -359,7 +360,7 @@ Nuevos, propios de esta migración:
 - **No engordar el SDK** durante la migración (D11). Regla del dos: ningún puerto ni método nuevo sin un segundo consumidor real esperándolo.
 - **No accounts/leaf nodes el día uno** (D6). La federación tiene su propio trigger.
 - **No HA/RAFT de 3 nodos** sin un SLA que lo pida.
-- **No colas compartidas entre proyectos** — nunca. Interop de operaciones sí; si un proyecto necesita invocar trabajo en otro: kicker HTTP con HMAC en el borde (y, post-migración, import/export de subjects como opción futura).
+- **No colas compartidas entre proyectos** — nunca. Interop de operaciones sí; si un proyecto necesita invocar trabajo en otro: una fachada HTTP con HMAC en el borde (retirada del paquete hasta que haga falta — spec §8.3) o, post-migración, import/export de subjects.
 - **No construir F4–F6 sobre Redis en paralelo** "por si acaso" — no pagar las dos facturas.
 - **No relajar la idempotencia de handlers** porque el broker tenga dedup (D4).
 - **No reconstruir un framework de tasks** (D14) — el receiver es una librería fina cuya config viene del contrato; si le crecen hooks genéricos o abstracciones "por si acaso", regla del dos.
